fix bugs in queue.c and gateway two theads read queue problem

+++ SensorNetwork_Qt/config.h
@@ -12,7 +12,7 @@ typedef uint16_t sensor_id_t;
 typedef uint16_t room_id_t;
 typedef int16_t sensor_value_t;     // temp in 0.1C, e.g. 18.3C is stored as 183
 typedef time_t sensor_ts_t;         // UTC timestamp as returned by time()
-typedef int16_t sensor_data_statu_t;
+typedef uint8_t sensor_data_statu_t;
 typedef struct {   // lab 6
 	sensor_id_t id;
     sensor_value_t tmp;
